# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SEAPEDIA API
  description: >
    Multi-role marketplace API (Guest, Buyer, Seller, Driver, Admin).
    Authorization is enforced server-side against a session's **active role**,
    not the full set of roles a user owns — see the `security` note on each
    protected operation below. All error responses share one envelope:
    `{ "error": { "code": "SOME_CODE", "message": "human readable" } }`.
  version: "1.0.0"
servers:
  - url: /api
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Session token returned by POST /auth/login, sent as
        `Authorization: Bearer <token>`. Sessions expire 7 days after login;
        logout deletes the session row immediately (true invalidation, not a
        client-side-only forget).
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
      required: [error]
...
    PublicProduct:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        price: { type: integer }
        stock: { type: integer }
        imageUrl:
          type: string
          nullable: true
          description: "Product photo URL — absolute (https://) or a same-origin path (/product-images/...)."
        isDeleted: { type: boolean }
        createdAt: { type: string, format: date-time }
...
  responses:
    Unauthenticated:
      description: Missing/invalid/expired bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { error: { code: UNAUTHENTICATED, message: "Missing bearer token" } }
    Forbidden:
      description: Active role missing or not permitted for this action (RBAC — checked server-side against the session's activeRole, never the client-claimed route).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { error: { code: WRONG_ROLE, message: "Active role not permitted for this action" } }
...
    NotFound:
      description: Resource not found (or not owned by the caller — cross-user access collapses to 404, not 403, to avoid confirming existence).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
paths:
...
  /products:
    get:
      tags: [catalog]
      summary: Public product listing (guest-accessible)
      security: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Case-insensitive substring match on product name.
        - name: storeId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items: { $ref: '#/components/schemas/PublicProduct' }
  /products/{id}:
    get:
      tags: [catalog]
      summary: Public product detail
      security: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { product: { $ref: '#/components/schemas/PublicProduct' } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /seller/products:
    get:
      tags: [seller]
      summary: List the caller's own (non-deleted) products
      security: [{ bearerAuth: [] }]
      description: Requires activeRole=SELLER.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items: { $ref: '#/components/schemas/PublicProduct' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [seller]
      summary: Create a product under the caller's store
      security: [{ bearerAuth: [] }]
      description: Requires activeRole=SELLER.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, stock]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                description: { type: string, maxLength: 2000 }
                price: { type: integer, minimum: 0 }
                stock: { type: integer, minimum: 0 }
                imageUrl:
                  type: string
                  nullable: true
                  maxLength: 500
                  description: "Must start with https://, http://, or /. Omit or set to null to leave/clear it."
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties: { product: { $ref: '#/components/schemas/PublicProduct' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: STORE_NOT_FOUND
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /seller/products/{id}:
    put:
      tags: [seller]
      summary: Update a product owned by the caller's store
      security: [{ bearerAuth: [] }]
      description: Requires activeRole=SELLER.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, stock]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                description: { type: string, maxLength: 2000 }
                price: { type: integer, minimum: 0 }
                stock: { type: integer, minimum: 0 }
                imageUrl:
                  type: string
                  nullable: true
                  maxLength: 500
                  description: "Must start with https://, http://, or /. Set to null to clear it."
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { product: { $ref: '#/components/schemas/PublicProduct' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403':
          description: WRONG_ROLE, or NOT_OWNER if the product belongs to a different seller's store.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [seller]
      summary: Soft-delete a product owned by the caller's store
      description: >
        Sets isDeleted=true (terminal — a soft-deleted product 404s on any
        further update/delete). Requires activeRole=SELLER.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403':
          description: WRONG_ROLE, or NOT_OWNER if the product belongs to a different seller's store.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/products:
    get:
      tags: [admin]
      summary: List all products (including soft-deleted) with their store name
      security: [{ bearerAuth: [] }]
      description: Requires activeRole=ADMIN.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        price: { type: integer }
                        stock: { type: integer }
                        imageUrl: { type: string, nullable: true }
                        isDeleted: { type: boolean }
                        storeName: { type: string }
                        createdAt: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
security:
  - bearerAuth: []